+++ tests/Functional/Tree/Multi/RootTest.php
@@ -19,7 +19,7 @@ class RootTest extends AbstractFunctionalTreeTestCase
     }
 
     #[Test]
-    public function createRoot(): void
+    public function createRootModel(): void
     {
         /** @var MultiCategory $model */
         $model = static::model(['title' => 'root node']);